Strict API Usage
From the looks of it, only strings can be the keys for entries in dictionaries, so I'm adding a validation step to ensure it is the correct type there.

I also moved usage of Buffer.slice() over to Buffer.subarray(), as I want to move to the more general Uint8Array class instead, anyways.

https://stackoverflow.com/questions/63162981/node-js-difference-between-buffer-slice-and-buffer-subarray