UCS-2 Debug Fix!
I was just wrapping the Buffer incorrectly! Heck yeah! So it's not the actual text format itself. This properly removes the need of the Node `Buffer` API.

I debugged it with the output from `plutil`, where I verified that the text it received indeed should be the same, because it's the same name present across each of the entries.

```sh
plutil -p ./test/utf16_chinese.plist
```

So from what I can tell, yes UTF-16 appears to be valid in supporting that of UCS-2. My hopes were that it was a superset, rather than a subset.

https://stackoverflow.com/questions/70720068/decode-base64-with-ucs-2-encoding
https://stackoverflow.com/questions/21797299/convert-base64-string-to-arraybuffer/41106346#41106346
https://groups.google.com/g/microsoft.public.vc.language/c/J6xhiU-NEjs?pli=1