Node Test Runner
Looks like it's just an issue with some recent updates to `--watch` mode in combination with the `--test` flag. I ran the same `node --test --watch ./demo.test.js` on a demo file, and it didn't have any tests in it, but the canceled error was still showing up. So it might have something to do with how they setup the usage with ESM, I have a feeling? Not completely sure. I went over to my test script in NBTify and ran it in the same current Node version, and it looks to be happening there now as well. So I feel like it's related to Node's watch mode specifically, rather than the test flag or the script itself.