Full TS Test + Nice and Tidy Config / Dependencies
Is it too opinionated of me to do this, when contributing to repos from a while back? Bringing projects up to the most modern tooling seems like the most helpful way to go, at least from my perspective I'd be really happy if someone did that for me down the line, so I don't think I'd mind. Not everyone may want that though, I'm not totally sure.