Browser Support!
Now the module doesn't use any Node APIs! So it means you can run it in the browser now as well.

For the sake of it, I'm curious how this could work in conjunction with converting back and forth between NBT haha.

I also need to write a writer for this project, this one only allows for reading plist files. I wanted to modernize the codebase like this, before moving on to trying that though.