File Async/Sync Test Removal
Now that the fs-related features are abstracted out of the library, it doesn't make sense to test both the async and sync usage of the file, because it is just to run the test now, instead of in the main module itself. I think my goal with it is that users can bring their own methods of getting the file content themselves, rather than having the package do that itself. I've gone that route with NBTify as well, since it's meant for Node and the browser.