Swap Bytes Mutation
Since this function modifies the buffer itself, I'm going to remove it's usage of return values. If it cloned the buffer and returned a new one, I would still use return values/reassignment, but I think this makes more sense instead.

Also just realized that it looks like it might be using `utf-16be` instead of `utf-16le`. I tried changing it out for that, and removing the usage of `swapBytes()`, and it worked! I'll do that next.