readUInt Dedupe Calls
Not sure why two of the same calls are in this if block, for both cases, but all of the tests are the same with removing them, and the code is identical for both, so unless this did something different previously, I don't think this check does anything currently. This could be previously set up for a feature that we don't have a file test setup for though, so that could be possibly something to look out for down the road.